requirements.txt
discord.py
pystyle
requests
